BIKE: Fix compiler complaints

+++ bike/src/bike_sampling_internal.h
@@ -45,8 +45,8 @@ extern "C" {
 // Value used to denote an invalid index for ther error vector.
 #define IDX_INVALID_VAL (0xffffffff)
 
-void secure_set_bits_port(pad_r_t *r, size_t first_pos, const idx_t *wlist,
-			  size_t w_size);
+void secure_set_bits_port(pad_r_t *r, uint32_t first_pos, const idx_t *wlist,
+			  uint32_t w_size);
 
 #if defined(UNIFORM_SAMPLING)
 void sample_error_vec_indices_port(idx_t *out, struct lc_hash_ctx *prf_state);
@@ -54,10 +54,10 @@ void sample_error_vec_indices_port(idx_t *out, struct lc_hash_ctx *prf_state);
 
 #if defined(X86_64)
 void secure_set_bits_avx2(pad_r_t *r, size_t first_pos, const idx_t *wlist,
-			  size_t w_size);
+			  uint32_t w_size);
 
 void secure_set_bits_avx512(pad_r_t *r, size_t first_pos, const idx_t *wlist,
-			    size_t w_size);
+			    uint32_t w_size);
 
 #if defined(UNIFORM_SAMPLING)
 void sample_error_vec_indices_avx2(idx_t *out, struct lc_hash_ctx *prf_state);
@@ -66,8 +66,8 @@ void sample_error_vec_indices_avx512(idx_t *out, struct lc_hash_ctx *prf_state);
 #endif
 
 typedef struct sampling_ctx_st {
-	void (*secure_set_bits)(pad_r_t *r, size_t first_pos,
-				const idx_t *wlist, size_t w_size);
+	void (*secure_set_bits)(pad_r_t *r, uint32_t first_pos,
+				const idx_t *wlist, uint32_t w_size);
 
 #if defined(UNIFORM_SAMPLING)
 	void (*sample_error_vec_indices)(idx_t *out,
